Fix für Plot Option
